unable remove current user from permissions

+++ dlg_addLayer.py
@@ -217,11 +217,10 @@ class AddLayerDialog(QtWidgets.QDialog, FORM_CLASS):
             self.pushButton_save.clicked.connect(lambda: self.askForMapPermissionChanges(layerName, usersDict, "layers"))
             self.pushButton_addRead.clicked.connect(lambda:  self.checkAddedItemDuplicity("read"))
             self.pushButton_addWrite.clicked.connect(lambda: self.setWritePermissionList())
-            self.pushButton_removeRead.clicked.connect(lambda: self.removeWritePermissionList(usersDictReversed))
-            self.pushButton_removeWrite.clicked.connect(lambda: self.listWidget_write.removeItemWidget(self.listWidget_write.takeItem(self.listWidget_write.currentRow())))
-            self.permissionsConnected = True
-       
-        print(self.listWidget_read.currentRow())
+            self.pushButton_removeRead.clicked.connect(lambda: self.removeReadPermissionList(usersDictReversed))
+            self.pushButton_removeWrite.clicked.connect(lambda: self.removeWritePermissionList(usersDictReversed))
+            self.permissionsConnected = True      
+  
 
     def callDeleteLayer(self, layers, layerNames):
         items = list()
@@ -538,8 +537,7 @@ class AddLayerDialog(QtWidgets.QDialog, FORM_CLASS):
         if checked == "0":
             checked = False
         if checked == "1":
-            checked = True
-        #self.loadLayersThread(checked)
+            checked = True      
         self.getLayers.emit(checked)
 
         if response.status_code == 200:
@@ -753,12 +751,17 @@ class AddLayerDialog(QtWidgets.QDialog, FORM_CLASS):
                 else:
                     QMessageBox.information(None, "Error", "Permissions was not saved for layer: " + str(failed).replace("[","").replace("]",""))                
                     
-    def removeWritePermissionList(self, usersDictReversed):  
-        # print(self.listWidget_read.currentRow())
-        # if usersDictReversed[self.laymanUsername] == self.listWidget_read.currentRow() or usersDictReversed[self.laymanUsername] == self.listWidget_write.currentRow():
-        #     return
+    def removeReadPermissionList(self, usersDictReversed):       
+        if usersDictReversed[self.laymanUsername] == self.listWidget_read.currentItem().text() or usersDictReversed[self.laymanUsername] == self.listWidget_write.currentRow():
+            self.utils.showQgisBar(["Není možné odebrat aktuálního uživatele.","Unable to remove current user."], Qgis.Warning)
+            return
         self.deleteItem(self.listWidget_read.currentItem().text())
         self.listWidget_read.removeItemWidget(self.listWidget_read.takeItem(self.listWidget_read.currentRow()))
+    def removeWritePermissionList(self, usersDictReversed):
+        if usersDictReversed[self.laymanUsername] == self.listWidget_write.currentItem().text():
+            self.utils.showQgisBar(["Není možné odebrat aktuálního uživatele.","Unable to remove current user."], Qgis.Warning)
+            return        
+        self.listWidget_write.removeItemWidget(self.listWidget_write.takeItem(self.listWidget_write.currentRow()))
     def deleteItem(self, itemName):
         items_list = self.listWidget_write.findItems(itemName, Qt.MatchExactly)
         for item in items_list:
